Revert "chore: Rebase to 3.36.4"
This reverts commit afeacee882b95ac0cb185461376636e20655036e.
