--- debian/patches/screen-cast-src-Fix-signedness-of-timestamp-field.patch
@@ -1,26 +1,0 @@
-From: =?utf-8?q?Jonas_=C3=85dahl?= <jadahl@gmail.com>
-Date: Fri, 3 Jul 2020 23:50:28 +0200
-Subject: screen-cast/src: Fix signedness of timestamp field
-
-https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1351
-
-Origin: https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1351
-Bug-GNOME: https://gitlab.gnome.org/GNOME/mutter/-/issues/1323
-Applied-Upstream: 3.36.5
----
- src/backends/meta-screen-cast-stream-src.c | 2 +-
- 1 file changed, 1 insertion(+), 1 deletion(-)
-
-diff --git a/src/backends/meta-screen-cast-stream-src.c b/src/backends/meta-screen-cast-stream-src.c
-index 4b79119..9921681 100644
---- a/src/backends/meta-screen-cast-stream-src.c
-+++ b/src/backends/meta-screen-cast-stream-src.c
-@@ -91,7 +91,7 @@ typedef struct _MetaScreenCastStreamSrcPrivate
-   struct spa_video_info_raw video_format;
-   int video_stride;
- 
--  uint64_t last_frame_timestamp_us;
-+  int64_t last_frame_timestamp_us;
- 
-   GHashTable *dmabuf_handles;
-
