--- debian/patches/screen-cast-src-Use-G_USEC_PER_SEC-instead-of-1000000.patch
@@ -1,29 +1,0 @@
-From: =?utf-8?q?Jonas_=C3=85dahl?= <jadahl@gmail.com>
-Date: Wed, 8 Jul 2020 15:08:23 +0200
-Subject: screen-cast/src: Use G_USEC_PER_SEC instead of 1000000
-
-https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1351
-
-Origin: https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1351
-Bug-GNOME: https://gitlab.gnome.org/GNOME/mutter/-/issues/1323
-Applied-Upstream: 3.36.5
----
- src/backends/meta-screen-cast-stream-src.c | 5 +++--
- 1 file changed, 3 insertions(+), 2 deletions(-)
-
-diff --git a/src/backends/meta-screen-cast-stream-src.c b/src/backends/meta-screen-cast-stream-src.c
-index 8412642..070c67a 100644
---- a/src/backends/meta-screen-cast-stream-src.c
-+++ b/src/backends/meta-screen-cast-stream-src.c
-@@ -515,8 +515,9 @@ meta_screen_cast_stream_src_maybe_record_frame (MetaScreenCastStreamSrc  *src,
-       int64_t min_interval_us;
-       int64_t time_since_last_frame_us;
- 
--      min_interval_us = ((1000000 * priv->video_format.max_framerate.denom) /
--                         priv->video_format.max_framerate.num);
-+      min_interval_us =
-+        ((G_USEC_PER_SEC * priv->video_format.max_framerate.denom) /
-+         priv->video_format.max_framerate.num);
- 
-       time_since_last_frame_us = now_us - priv->last_frame_timestamp_us;
-       if (time_since_last_frame_us < min_interval_us)
