--- debian/patches/screen-cast-window-stream-src-Fix-indentation.patch
@@ -1,28 +1,0 @@
-From: =?utf-8?q?Jonas_=C3=85dahl?= <jadahl@gmail.com>
-Date: Fri, 3 Jul 2020 16:46:44 +0200
-Subject: screen-cast/window-stream-src: Fix indentation
-
-https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1351
-
-Origin: https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1351
-Bug-GNOME: https://gitlab.gnome.org/GNOME/mutter/-/issues/1323
-Applied-Upstream: 3.36.5
----
- src/backends/meta-screen-cast-window-stream-src.c | 4 ++--
- 1 file changed, 2 insertions(+), 2 deletions(-)
-
-diff --git a/src/backends/meta-screen-cast-window-stream-src.c b/src/backends/meta-screen-cast-window-stream-src.c
-index 5d1bfde..3f141d2 100644
---- a/src/backends/meta-screen-cast-window-stream-src.c
-+++ b/src/backends/meta-screen-cast-window-stream-src.c
-@@ -477,8 +477,8 @@ meta_screen_cast_window_stream_src_record_to_framebuffer (MetaScreenCastStreamSr
-   stream_rect.height = get_stream_height (window_src);
- 
-   if (!meta_screen_cast_window_blit_to_framebuffer (window_src->screen_cast_window,
--                                                     &stream_rect,
--                                                     framebuffer))
-+                                                    &stream_rect,
-+                                                    framebuffer))
-     return FALSE;
- 
-   stream = meta_screen_cast_stream_src_get_stream (src);
