+++ src/backends/native/meta-renderer-native.c
@@ -2146,7 +2146,6 @@ meta_renderer_native_create_dma_buf (CoglRenderer  *cogl_renderer,
     case META_RENDERER_NATIVE_MODE_GBM:
       {
         CoglFramebuffer *dmabuf_fb;
-        CoglDmaBufHandle *dmabuf_handle;
         struct gbm_bo *new_bo;
         int dmabuf_fd = -1;
 
@@ -2183,11 +2182,8 @@ meta_renderer_native_create_dma_buf (CoglRenderer  *cogl_renderer,
         if (!dmabuf_fb)
           return NULL;
 
-        dmabuf_handle =
-          cogl_dma_buf_handle_new (dmabuf_fb, dmabuf_fd, new_bo,
-                                   (GDestroyNotify) gbm_bo_destroy);
-        cogl_object_unref (dmabuf_fb);
-        return dmabuf_handle;
+        return cogl_dma_buf_handle_new (dmabuf_fb, dmabuf_fd, new_bo,
+                                        (GDestroyNotify) gbm_bo_destroy);
       }
       break;
 #ifdef HAVE_EGL_DEVICE
