+++ src/meson.build
@@ -495,22 +495,7 @@ if have_wayland
     'wayland/meta-wayland-cursor-surface.h',
     'wayland/meta-wayland-data-device.c',
     'wayland/meta-wayland-data-device.h',
-    'wayland/meta-wayland-data-device-primary.c',
-    'wayland/meta-wayland-data-device-primary.h',
-    'wayland/meta-wayland-data-device-primary-legacy.c',
-    'wayland/meta-wayland-data-device-primary-legacy.h',
-    'wayland/meta-wayland-data-offer.c',
-    'wayland/meta-wayland-data-offer.h',
-    'wayland/meta-wayland-data-offer-primary.c',
-    'wayland/meta-wayland-data-offer-primary.h',
-    'wayland/meta-wayland-data-offer-primary-legacy.c',
-    'wayland/meta-wayland-data-offer-primary-legacy.h',
-    'wayland/meta-wayland-data-source.c',
-    'wayland/meta-wayland-data-source.h',
-    'wayland/meta-wayland-data-source-primary.c',
-    'wayland/meta-wayland-data-source-primary.h',
-    'wayland/meta-wayland-data-source-primary-legacy.c',
-    'wayland/meta-wayland-data-source-primary-legacy.h',
+    'wayland/meta-wayland-data-device-private.h',
     'wayland/meta-wayland-dma-buf.c',
     'wayland/meta-wayland-dma-buf.h',
     'wayland/meta-wayland-dnd-surface.c',
@@ -805,7 +790,6 @@ if have_wayland
     ['linux-dmabuf', 'unstable', 'v1', ],
     ['pointer-constraints', 'unstable', 'v1', ],
     ['pointer-gestures', 'unstable', 'v1', ],
-    ['primary-selection', 'unstable', 'v1', ],
     ['relative-pointer', 'unstable', 'v1', ],
     ['tablet', 'unstable', 'v2', ],
     ['text-input', 'unstable', 'v3', ],
