+++ src/wayland/meta-wayland-data-device.c
@@ -25,6 +25,7 @@
 #include "config.h"
 
 #include "wayland/meta-wayland-data-device.h"
+#include "wayland/meta-wayland-data-device-private.h"
 
 #include <gio/gunixoutputstream.h>
 #include <glib-unix.h>
@@ -42,16 +43,70 @@
 #include "wayland/meta-wayland-private.h"
 #include "wayland/meta-wayland-seat.h"
 
+#include "gtk-primary-selection-server-protocol.h"
+
 #define ROOTWINDOW_DROP_MIME "application/x-rootwindow-drop"
+
+#define ALL_ACTIONS (WL_DATA_DEVICE_MANAGER_DND_ACTION_COPY | \
+                     WL_DATA_DEVICE_MANAGER_DND_ACTION_MOVE | \
+                     WL_DATA_DEVICE_MANAGER_DND_ACTION_ASK)
+
+struct _MetaWaylandDataOffer
+{
+  struct wl_resource *resource;
+  MetaWaylandDataSource *source;
+  struct wl_listener source_destroy_listener;
+  gboolean accepted;
+  gboolean action_sent;
+  uint32_t dnd_actions;
+  enum wl_data_device_manager_dnd_action preferred_dnd_action;
+  MetaSelectionType selection_type;
+};
+
+typedef struct _MetaWaylandDataSourcePrivate
+{
+  struct wl_resource *resource;
+  MetaWaylandDataOffer *offer;
+  struct wl_array mime_types;
+  gboolean has_target;
+  uint32_t dnd_actions;
+  enum wl_data_device_manager_dnd_action user_dnd_action;
+  enum wl_data_device_manager_dnd_action current_dnd_action;
+  MetaWaylandSeat *seat;
+  guint actions_set : 1;
+  guint in_ask : 1;
+  guint drop_performed : 1;
+} MetaWaylandDataSourcePrivate;
+
+typedef struct _MetaWaylandDataSourcePrimary
+{
+  MetaWaylandDataSource parent;
+} MetaWaylandDataSourcePrimary;
+
+G_DEFINE_TYPE_WITH_PRIVATE (MetaWaylandDataSource, meta_wayland_data_source,
+                            G_TYPE_OBJECT);
+G_DEFINE_TYPE (MetaWaylandDataSourcePrimary, meta_wayland_data_source_primary,
+               META_TYPE_WAYLAND_DATA_SOURCE);
 
 static void unset_selection_source (MetaWaylandDataDevice *data_device,
                                     MetaSelectionType      selection_type);
+
+static MetaWaylandDataSource *
+meta_wayland_data_source_new (struct wl_resource *resource);
+static MetaWaylandDataSource *
+meta_wayland_data_source_primary_new (struct wl_resource *resource);
+
+static void
+meta_wayland_source_cancel (MetaWaylandDataSource *source);
 
 static void
 drag_grab_data_source_destroyed (gpointer data, GObject *where_the_object_was);
 
 static struct wl_resource * create_and_send_clipboard_offer (MetaWaylandDataDevice *data_device,
                                                              struct wl_resource    *target);
+static struct wl_resource * create_and_send_primary_offer   (MetaWaylandDataDevice *data_device,
+                                                             struct wl_resource    *target);
+static struct wl_resource * meta_wayland_data_source_get_resource (MetaWaylandDataSource *source);
 
 static void
 move_resources (struct wl_list *destination,
@@ -83,6 +138,347 @@ unbind_resource (struct wl_resource *resource)
   wl_list_remove (wl_resource_get_link (resource));
 }
 
+static gboolean
+meta_wayland_source_get_in_ask (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  return priv->in_ask;
+}
+
+static void
+meta_wayland_source_update_in_ask (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  priv->in_ask =
+    priv->current_dnd_action == WL_DATA_DEVICE_MANAGER_DND_ACTION_ASK;
+}
+
+static enum wl_data_device_manager_dnd_action
+data_offer_choose_action (MetaWaylandDataOffer *offer)
+{
+  MetaWaylandDataSource *source = offer->source;
+  uint32_t actions, user_action, available_actions;
+
+  if (wl_resource_get_version (offer->resource) <
+      WL_DATA_OFFER_ACTION_SINCE_VERSION)
+    return WL_DATA_DEVICE_MANAGER_DND_ACTION_COPY;
+
+  actions = meta_wayland_data_source_get_actions (source);
+  user_action = meta_wayland_data_source_get_user_action (source);
+
+  available_actions = actions & offer->dnd_actions;
+
+  if (!available_actions)
+    return WL_DATA_DEVICE_MANAGER_DND_ACTION_NONE;
+
+  /* If the user is forcing an action, go for it */
+  if ((user_action & available_actions) != 0)
+    return user_action;
+
+  /* If the dest side has a preferred DnD action, use it */
+  if ((offer->preferred_dnd_action & available_actions) != 0)
+    return offer->preferred_dnd_action;
+
+  /* Use the first found action, in bit order */
+  return 1 << (ffs (available_actions) - 1);
+}
+
+static void
+data_offer_update_action (MetaWaylandDataOffer *offer)
+{
+  enum wl_data_device_manager_dnd_action current_action, action;
+  MetaWaylandDataSource *source;
+
+  if (!offer->source)
+    return;
+
+  source = offer->source;
+  current_action = meta_wayland_data_source_get_current_action (source);
+  action = data_offer_choose_action (offer);
+
+  if (current_action == action)
+    return;
+
+  meta_wayland_data_source_set_current_action (source, action);
+
+  if (!meta_wayland_source_get_in_ask (source) &&
+      wl_resource_get_version (offer->resource) >=
+      WL_DATA_OFFER_ACTION_SINCE_VERSION)
+    {
+      wl_data_offer_send_action (offer->resource, action);
+      offer->action_sent = TRUE;
+    }
+}
+
+static struct wl_resource *
+meta_wayland_data_source_get_resource (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  return priv->resource;
+}
+
+static void
+meta_wayland_data_source_set_resource (MetaWaylandDataSource *source,
+                                       struct wl_resource    *resource)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  priv->resource = resource;
+}
+
+static void
+meta_wayland_data_source_target (MetaWaylandDataSource *source,
+                                 const char *mime_type)
+{
+  if (META_WAYLAND_DATA_SOURCE_GET_CLASS (source)->target)
+    META_WAYLAND_DATA_SOURCE_GET_CLASS (source)->target (source, mime_type);
+}
+
+void
+meta_wayland_data_source_send (MetaWaylandDataSource *source,
+                               const char *mime_type,
+                               int fd)
+{
+  META_WAYLAND_DATA_SOURCE_GET_CLASS (source)->send (source, mime_type, fd);
+}
+
+gboolean
+meta_wayland_data_source_has_target (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  return priv->has_target;
+}
+
+static void
+meta_wayland_data_source_set_seat (MetaWaylandDataSource *source,
+                                   MetaWaylandSeat       *seat)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  priv->seat = seat;
+}
+
+static MetaWaylandSeat *
+meta_wayland_data_source_get_seat (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  return priv->seat;
+}
+
+void
+meta_wayland_data_source_set_has_target (MetaWaylandDataSource *source,
+                                         gboolean has_target)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  priv->has_target = has_target;
+}
+
+struct wl_array *
+meta_wayland_data_source_get_mime_types (const MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private ((MetaWaylandDataSource *)source);
+
+  return &priv->mime_types;
+}
+
+void
+meta_wayland_data_source_cancel (MetaWaylandDataSource *source)
+{
+  META_WAYLAND_DATA_SOURCE_GET_CLASS (source)->cancel (source);
+}
+
+uint32_t
+meta_wayland_data_source_get_actions (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  return priv->dnd_actions;
+}
+
+enum wl_data_device_manager_dnd_action
+meta_wayland_data_source_get_user_action (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  if (!priv->seat)
+    return WL_DATA_DEVICE_MANAGER_DND_ACTION_NONE;
+
+  return priv->user_dnd_action;
+}
+
+enum wl_data_device_manager_dnd_action
+meta_wayland_data_source_get_current_action (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  return priv->current_dnd_action;
+}
+
+static void
+meta_wayland_data_source_set_current_offer (MetaWaylandDataSource *source,
+                                            MetaWaylandDataOffer  *offer)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  priv->offer = offer;
+}
+
+static MetaWaylandDataOffer *
+meta_wayland_data_source_get_current_offer (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  return priv->offer;
+}
+
+void
+meta_wayland_data_source_set_current_action (MetaWaylandDataSource                  *source,
+                                             enum wl_data_device_manager_dnd_action  action)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  if (priv->current_dnd_action == action)
+    return;
+
+  priv->current_dnd_action = action;
+
+  if (!meta_wayland_source_get_in_ask (source))
+    META_WAYLAND_DATA_SOURCE_GET_CLASS (source)->action (source, action);
+}
+
+void
+meta_wayland_data_source_set_actions (MetaWaylandDataSource *source,
+                                      uint32_t               dnd_actions)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  priv->dnd_actions = dnd_actions;
+  priv->actions_set = TRUE;
+}
+
+void
+meta_wayland_data_source_set_user_action (MetaWaylandDataSource *source,
+                                          uint32_t               action)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+  MetaWaylandDataOffer *offer;
+
+  if (priv->user_dnd_action == action)
+    return;
+
+  priv->user_dnd_action = action;
+  offer = meta_wayland_data_source_get_current_offer (source);
+
+  if (offer)
+    data_offer_update_action (offer);
+}
+
+static gboolean
+meta_wayland_data_source_get_drop_performed (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  return priv->drop_performed;
+}
+
+static void
+data_offer_accept (struct wl_client *client,
+                   struct wl_resource *resource,
+                   guint32 serial,
+                   const char *mime_type)
+{
+  MetaWaylandDataOffer *offer = wl_resource_get_user_data (resource);
+
+  /* FIXME: Check that client is currently focused by the input
+   * device that is currently dragging this data source.  Should
+   * this be a wl_data_device request? */
+
+  if (offer->source)
+    {
+      meta_wayland_data_source_target (offer->source, mime_type);
+      meta_wayland_data_source_set_has_target (offer->source,
+                                               mime_type != NULL);
+    }
+
+  offer->accepted = mime_type != NULL;
+}
+
+static void
+transfer_cb (MetaSelection *selection,
+             GAsyncResult  *res,
+             GOutputStream *stream)
+{
+  GError *error = NULL;
+
+  if (!meta_selection_transfer_finish (selection, res, &error))
+    {
+      g_warning ("Could not fetch selection data: %s", error->message);
+      g_error_free (error);
+    }
+
+  g_output_stream_close (stream, NULL, NULL);
+}
+
+static void
+data_offer_receive (struct wl_client *client, struct wl_resource *resource,
+                    const char *mime_type, int32_t fd)
+{
+  MetaWaylandDataOffer *offer = wl_resource_get_user_data (resource);
+  MetaDisplay *display = meta_get_display ();
+  MetaSelectionType selection_type;
+  GList *mime_types;
+  gboolean found;
+
+  selection_type = offer->selection_type;
+  mime_types = meta_selection_get_mimetypes (meta_display_get_selection (display),
+                                             selection_type);
+  found = g_list_find_custom (mime_types, mime_type, (GCompareFunc) g_strcmp0) != NULL;
+  g_list_free_full (mime_types, g_free);
+
+  if (found)
+    {
+      GOutputStream *stream;
+
+      stream = g_unix_output_stream_new (fd, TRUE);
+      meta_selection_transfer_async (meta_display_get_selection (display),
+                                     selection_type,
+                                     mime_type,
+                                     -1,
+                                     stream,
+                                     NULL,
+                                     (GAsyncReadyCallback) transfer_cb,
+                                     stream);
+    }
+  else
+    {
+      close (fd);
+    }
+}
+
 static void
 default_destructor (struct wl_client   *client,
                     struct wl_resource *resource)
@@ -90,30 +486,284 @@ default_destructor (struct wl_client   *client,
   wl_resource_destroy (resource);
 }
 
+static void
+data_offer_finish (struct wl_client   *client,
+		   struct wl_resource *resource)
+{
+  MetaWaylandDataOffer *offer = wl_resource_get_user_data (resource);
+  enum wl_data_device_manager_dnd_action current_action;
+
+  if (!offer->source ||
+      offer != meta_wayland_data_source_get_current_offer (offer->source))
+    return;
+
+  if (!offer->accepted || !offer->action_sent)
+    {
+      wl_resource_post_error (offer->resource,
+                              WL_DATA_OFFER_ERROR_INVALID_FINISH,
+                              "premature finish request");
+      return;
+    }
+
+  current_action = meta_wayland_data_source_get_current_action (offer->source);
+
+  if (current_action == WL_DATA_DEVICE_MANAGER_DND_ACTION_NONE ||
+      current_action == WL_DATA_DEVICE_MANAGER_DND_ACTION_ASK)
+    {
+      wl_resource_post_error (offer->resource,
+                              WL_DATA_OFFER_ERROR_INVALID_OFFER,
+                              "offer finished with an invalid action");
+      return;
+    }
+
+  meta_wayland_data_source_notify_finish (offer->source);
+}
+
+static void
+data_offer_set_actions (struct wl_client   *client,
+                        struct wl_resource *resource,
+                        uint32_t            dnd_actions,
+                        uint32_t            preferred_action)
+{
+  MetaWaylandDataOffer *offer = wl_resource_get_user_data (resource);
+
+  if (dnd_actions & ~(ALL_ACTIONS))
+    {
+      wl_resource_post_error (offer->resource,
+                              WL_DATA_OFFER_ERROR_INVALID_ACTION_MASK,
+                              "invalid actions mask %x", dnd_actions);
+      return;
+    }
+
+  if (preferred_action &&
+      (!(preferred_action & dnd_actions) ||
+       __builtin_popcount (preferred_action) > 1))
+    {
+      wl_resource_post_error (offer->resource,
+                              WL_DATA_OFFER_ERROR_INVALID_ACTION,
+                              "invalid action %x", preferred_action);
+      return;
+    }
+
+  offer->dnd_actions = dnd_actions;
+  offer->preferred_dnd_action = preferred_action;
+
+  data_offer_update_action (offer);
+}
+
+static const struct wl_data_offer_interface data_offer_interface = {
+  data_offer_accept,
+  data_offer_receive,
+  default_destructor,
+  data_offer_finish,
+  data_offer_set_actions,
+};
+
+static void
+primary_offer_receive (struct wl_client *client, struct wl_resource *resource,
+                       const char *mime_type, int32_t fd)
+{
+  MetaDisplay *display = meta_get_display ();
+  GOutputStream *stream;
+  GList *mime_types;
+  gboolean found;
+
+  mime_types = meta_selection_get_mimetypes (meta_display_get_selection (display),
+                                             META_SELECTION_PRIMARY);
+  found = g_list_find_custom (mime_types, mime_type, (GCompareFunc) g_strcmp0) != NULL;
+  g_list_free_full (mime_types, g_free);
+
+  if (!found)
+    {
+      close (fd);
+      return;
+    }
+
+  stream = g_unix_output_stream_new (fd, TRUE);
+  meta_selection_transfer_async (meta_display_get_selection (display),
+                                 META_SELECTION_PRIMARY,
+                                 mime_type,
+                                 -1,
+                                 stream,
+                                 NULL,
+                                 (GAsyncReadyCallback) transfer_cb,
+                                 stream);
+}
+
+static const struct gtk_primary_selection_offer_interface primary_offer_interface = {
+  primary_offer_receive,
+  default_destructor,
+};
+
+static void
+meta_wayland_data_source_notify_drop_performed (MetaWaylandDataSource *source)
+{
+  META_WAYLAND_DATA_SOURCE_GET_CLASS (source)->drop_performed (source);
+}
+
+void
+meta_wayland_data_source_notify_finish (MetaWaylandDataSource *source)
+{
+  META_WAYLAND_DATA_SOURCE_GET_CLASS (source)->drag_finished (source);
+}
+
+static void
+destroy_data_offer (struct wl_resource *resource)
+{
+  MetaWaylandDataOffer *offer = wl_resource_get_user_data (resource);
+  MetaWaylandSeat *seat;
+
+  if (offer->source)
+    {
+      seat = meta_wayland_data_source_get_seat (offer->source);
+
+      if (offer == meta_wayland_data_source_get_current_offer (offer->source))
+        {
+          if (seat->data_device.dnd_data_source == offer->source)
+            {
+              if (wl_resource_get_version (offer->resource) <
+                  WL_DATA_OFFER_ACTION_SINCE_VERSION)
+                meta_wayland_data_source_notify_finish (offer->source);
+              else if (meta_wayland_data_source_get_drop_performed (offer->source))
+                meta_wayland_source_cancel (offer->source);
+            }
+          else
+            {
+              meta_wayland_data_source_set_current_offer (offer->source, NULL);
+              meta_wayland_data_source_set_has_target (offer->source, FALSE);
+            }
+        }
+
+      g_object_remove_weak_pointer (G_OBJECT (offer->source),
+                                    (gpointer *)&offer->source);
+      offer->source = NULL;
+    }
+
+  meta_display_sync_wayland_input_focus (meta_get_display ());
+  g_slice_free (MetaWaylandDataOffer, offer);
+}
+
+static void
+destroy_primary_offer (struct wl_resource *resource)
+{
+  MetaWaylandDataOffer *offer = wl_resource_get_user_data (resource);
+
+  if (offer->source)
+    {
+      if (offer == meta_wayland_data_source_get_current_offer (offer->source))
+        {
+          meta_wayland_data_source_cancel (offer->source);
+          meta_wayland_data_source_set_current_offer (offer->source, NULL);
+        }
+
+      g_object_remove_weak_pointer (G_OBJECT (offer->source),
+                                    (gpointer *)&offer->source);
+      offer->source = NULL;
+    }
+
+  meta_display_sync_wayland_input_focus (meta_get_display ());
+  g_slice_free (MetaWaylandDataOffer, offer);
+}
+
 static struct wl_resource *
 create_and_send_dnd_offer (MetaWaylandDataSource *source,
                            struct wl_resource *target)
 {
-  MetaWaylandDataOffer *offer;
-  struct wl_array *mime_types;
-  struct wl_resource *resource;
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+  MetaWaylandDataOffer *offer = g_slice_new0 (MetaWaylandDataOffer);
   char **p;
 
-  offer = meta_wayland_data_offer_new (META_SELECTION_DND, source, target);
-  resource = meta_wayland_data_offer_get_resource (offer);
+  offer->selection_type = META_SELECTION_DND;
+  offer->source = source;
+  g_object_add_weak_pointer (G_OBJECT (source), (gpointer *)&offer->source);
+  offer->resource = wl_resource_create (wl_resource_get_client (target),
+                                        &wl_data_offer_interface,
+                                        wl_resource_get_version (target), 0);
+  wl_resource_set_implementation (offer->resource,
+                                  &data_offer_interface,
+                                  offer,
+                                  destroy_data_offer);
 
-  wl_data_device_send_data_offer (target, resource);
+  wl_data_device_send_data_offer (target, offer->resource);
 
-  mime_types = meta_wayland_data_source_get_mime_types (source);
+  wl_array_for_each (p, &priv->mime_types)
+    wl_data_offer_send_offer (offer->resource, *p);
 
-  wl_array_for_each (p, mime_types)
-    wl_data_offer_send_offer (resource, *p);
-
-  meta_wayland_data_offer_update_action (offer);
+  data_offer_update_action (offer);
   meta_wayland_data_source_set_current_offer (source, offer);
 
-  return resource;
+  return offer->resource;
 }
+
+static void
+data_source_offer (struct wl_client *client,
+                   struct wl_resource *resource, const char *type)
+{
+  MetaWaylandDataSource *source = wl_resource_get_user_data (resource);
+
+  if (!meta_wayland_data_source_add_mime_type (source, type))
+    wl_resource_post_no_memory (resource);
+}
+
+static void
+data_source_set_actions (struct wl_client   *client,
+                         struct wl_resource *resource,
+                         uint32_t            dnd_actions)
+{
+  MetaWaylandDataSource *source = wl_resource_get_user_data (resource);
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  if (priv->actions_set)
+    {
+      wl_resource_post_error (priv->resource,
+                              WL_DATA_SOURCE_ERROR_INVALID_ACTION_MASK,
+                              "cannot set actions more than once");
+      return;
+    }
+
+  if (dnd_actions & ~(ALL_ACTIONS))
+    {
+      wl_resource_post_error (priv->resource,
+                              WL_DATA_SOURCE_ERROR_INVALID_ACTION_MASK,
+                              "invalid actions mask %x", dnd_actions);
+      return;
+    }
+
+  if (meta_wayland_data_source_get_seat (source))
+    {
+      wl_resource_post_error (priv->resource,
+                              WL_DATA_SOURCE_ERROR_INVALID_ACTION_MASK,
+                              "invalid action change after "
+                              "wl_data_device.start_drag");
+      return;
+    }
+
+  meta_wayland_data_source_set_actions (source, dnd_actions);
+}
+
+static struct wl_data_source_interface data_source_interface = {
+  data_source_offer,
+  default_destructor,
+  data_source_set_actions
+};
+
+static void
+primary_source_offer (struct wl_client   *client,
+                      struct wl_resource *resource,
+                      const char         *type)
+{
+  MetaWaylandDataSource *source = wl_resource_get_user_data (resource);
+
+  if (!meta_wayland_data_source_add_mime_type (source, type))
+    wl_resource_post_no_memory (resource);
+}
+
+static struct gtk_primary_selection_source_interface primary_source_interface = {
+  primary_source_offer,
+  default_destructor,
+};
 
 struct _MetaWaylandDragGrab {
   MetaWaylandPointerGrab  generic;
@@ -215,7 +865,7 @@ meta_wayland_drag_source_fake_acceptance (MetaWaylandDataSource *source,
 {
   uint32_t actions, user_action, action = 0;
 
-  meta_wayland_data_source_get_actions (source, &actions);
+  actions = meta_wayland_data_source_get_actions (source);
   user_action = meta_wayland_data_source_get_user_action (source);
 
   /* Pick a suitable action */
@@ -454,7 +1104,7 @@ drag_grab_button (MetaWaylandPointerGrab *grab,
           meta_wayland_surface_drag_dest_drop (drag_grab->drag_focus);
           meta_wayland_data_source_notify_drop_performed (source);
 
-          meta_wayland_data_source_update_in_ask (source);
+          meta_wayland_source_update_in_ask (source);
           success = TRUE;
         }
       else if (!drag_grab->drag_focus && source &&
@@ -730,15 +1380,171 @@ selection_data_source_destroyed (gpointer data, GObject *object_was_here)
 }
 
 static void
+meta_wayland_source_send (MetaWaylandDataSource *source,
+                          const gchar           *mime_type,
+                          gint                   fd)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  wl_data_source_send_send (priv->resource, mime_type, fd);
+  close (fd);
+}
+
+static void
+meta_wayland_source_target (MetaWaylandDataSource *source,
+                            const gchar           *mime_type)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  wl_data_source_send_target (priv->resource, mime_type);
+}
+
+static void
+meta_wayland_source_cancel (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  if (!priv->resource)
+    return;
+
+  wl_data_source_send_cancelled (priv->resource);
+}
+
+static void
+meta_wayland_source_action (MetaWaylandDataSource                  *source,
+                            enum wl_data_device_manager_dnd_action  action)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  if (wl_resource_get_version (priv->resource) >=
+      WL_DATA_SOURCE_ACTION_SINCE_VERSION)
+    wl_data_source_send_action (priv->resource, action);
+}
+
+static void
+meta_wayland_source_drop_performed (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  if (wl_resource_get_version (priv->resource) >=
+      WL_DATA_SOURCE_DND_DROP_PERFORMED_SINCE_VERSION)
+    {
+      priv->drop_performed = TRUE;
+      wl_data_source_send_dnd_drop_performed (priv->resource);
+    }
+}
+
+static void
+meta_wayland_source_drag_finished (MetaWaylandDataSource *source)
+{
+  MetaWaylandCompositor *compositor = meta_wayland_compositor_get_default ();
+  MetaWaylandDataDevice *data_device = &compositor->seat->data_device;
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+  enum wl_data_device_manager_dnd_action action;
+
+  if (meta_wayland_source_get_in_ask (source))
+    {
+      action = meta_wayland_data_source_get_current_action (source);
+      meta_wayland_source_action (source, action);
+    }
+
+  if (wl_resource_get_version (priv->resource) >=
+      WL_DATA_SOURCE_DND_FINISHED_SINCE_VERSION)
+    wl_data_source_send_dnd_finished (priv->resource);
+
+  unset_selection_source (data_device, META_SELECTION_DND);
+}
+
+static void
+meta_wayland_source_finalize (GObject *object)
+{
+  MetaWaylandDataSource *source = META_WAYLAND_DATA_SOURCE (object);
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+  char **pos;
+
+  wl_array_for_each (pos, &priv->mime_types)
+    g_free (*pos);
+  wl_array_release (&priv->mime_types);
+
+  G_OBJECT_CLASS (meta_wayland_data_source_parent_class)->finalize (object);
+}
+
+static void
+meta_wayland_data_source_init (MetaWaylandDataSource *source)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  wl_array_init (&priv->mime_types);
+  priv->current_dnd_action = -1;
+  priv->drop_performed = FALSE;
+}
+
+static void
+meta_wayland_data_source_class_init (MetaWaylandDataSourceClass *klass)
+{
+  GObjectClass *object_class = G_OBJECT_CLASS (klass);
+
+  object_class->finalize = meta_wayland_source_finalize;
+
+  klass->send = meta_wayland_source_send;
+  klass->target = meta_wayland_source_target;
+  klass->cancel = meta_wayland_source_cancel;
+  klass->action = meta_wayland_source_action;
+  klass->drop_performed = meta_wayland_source_drop_performed;
+  klass->drag_finished = meta_wayland_source_drag_finished;
+}
+
+static void
+meta_wayland_data_source_primary_send (MetaWaylandDataSource *source,
+                                       const gchar           *mime_type,
+                                       gint                   fd)
+{
+  struct wl_resource *resource = meta_wayland_data_source_get_resource (source);
+
+  gtk_primary_selection_source_send_send (resource, mime_type, fd);
+  close (fd);
+}
+
+static void
+meta_wayland_data_source_primary_cancel (MetaWaylandDataSource *source)
+{
+  struct wl_resource *resource = meta_wayland_data_source_get_resource (source);
+
+  if (resource)
+    gtk_primary_selection_source_send_cancelled (resource);
+}
+
+static void
+meta_wayland_data_source_primary_init (MetaWaylandDataSourcePrimary *source_primary)
+{
+}
+
+static void
+meta_wayland_data_source_primary_class_init (MetaWaylandDataSourcePrimaryClass *klass)
+{
+  MetaWaylandDataSourceClass *data_source_class =
+    META_WAYLAND_DATA_SOURCE_CLASS (klass);
+
+  data_source_class->send = meta_wayland_data_source_primary_send;
+  data_source_class->cancel = meta_wayland_data_source_primary_cancel;
+}
+
+static void
 meta_wayland_drag_dest_focus_in (MetaWaylandDataDevice *data_device,
                                  MetaWaylandSurface    *surface,
                                  MetaWaylandDataOffer  *offer)
 {
   MetaWaylandDragGrab *grab = data_device->current_grab;
-  MetaWaylandDataSource *source;
   struct wl_display *display;
   struct wl_client *client;
-  struct wl_resource *resource;
   uint32_t source_actions;
   wl_fixed_t sx, sy;
 
@@ -752,21 +1558,18 @@ meta_wayland_drag_dest_focus_in (MetaWaylandDataDevice *data_device,
   wl_resource_add_destroy_listener (grab->drag_focus_data_device,
                                     &grab->drag_focus_listener);
 
-  resource = meta_wayland_data_offer_get_resource (offer);
-
-  if (wl_resource_get_version (resource) >=
+  if (wl_resource_get_version (offer->resource) >=
       WL_DATA_OFFER_SOURCE_ACTIONS_SINCE_VERSION)
     {
-      source = meta_wayland_data_offer_get_source (offer);
-      meta_wayland_data_source_get_actions (source, &source_actions);
-      wl_data_offer_send_source_actions (resource, source_actions);
+      source_actions = meta_wayland_data_source_get_actions (offer->source);
+      wl_data_offer_send_source_actions (offer->resource, source_actions);
     }
 
   meta_wayland_pointer_get_relative_coordinates (grab->generic.pointer,
                                                  surface, &sx, &sy);
   wl_data_device_send_enter (grab->drag_focus_data_device,
                              wl_display_next_serial (display),
-                             surface->resource, sx, sy, resource);
+                             surface->resource, sx, sy, offer->resource);
 }
 
 static void
@@ -918,6 +1721,7 @@ data_device_set_selection (struct wl_client *client,
 {
   MetaWaylandDataDevice *data_device = wl_resource_get_user_data (resource);
   MetaWaylandSeat *seat = wl_container_of (data_device, seat, data_device);
+  MetaWaylandDataSourcePrivate *priv;
   MetaWaylandDataSource *source;
 
   if (source_resource)
@@ -927,7 +1731,9 @@ data_device_set_selection (struct wl_client *client,
 
   if (source)
     {
-      if (meta_wayland_data_source_get_actions (source, NULL))
+      priv = meta_wayland_data_source_get_instance_private (source);
+
+      if (priv->actions_set)
         {
           wl_resource_post_error(source_resource,
                                  WL_DATA_SOURCE_ERROR_INVALID_SOURCE,
@@ -949,6 +1755,93 @@ static const struct wl_data_device_interface data_device_interface = {
   data_device_set_selection,
   default_destructor,
 };
+
+static void
+primary_source_destroyed (gpointer  data,
+                          GObject  *object_was_here)
+{
+  MetaWaylandDataDevice *data_device = data;
+
+  data_device->primary_data_source = NULL;
+  unset_selection_source (data_device, META_SELECTION_PRIMARY);
+}
+
+void
+meta_wayland_data_device_set_primary (MetaWaylandDataDevice *data_device,
+                                      MetaWaylandDataSource *source,
+                                      guint32                serial)
+{
+  MetaWaylandSeat *seat = wl_container_of (data_device, seat, data_device);
+  MetaSelectionSource *selection_source;
+
+  g_assert (!source || META_IS_WAYLAND_DATA_SOURCE_PRIMARY (source));
+
+  if (data_device->primary_data_source &&
+      data_device->primary_serial - serial < UINT32_MAX / 2)
+    return;
+
+  if (data_device->primary_data_source)
+    {
+      g_object_weak_unref (G_OBJECT (data_device->primary_data_source),
+                           primary_source_destroyed,
+                           data_device);
+    }
+
+  data_device->primary_data_source = source;
+  data_device->primary_serial = serial;
+
+  if (source)
+    {
+      meta_wayland_data_source_set_seat (source, seat);
+      g_object_weak_ref (G_OBJECT (source),
+                         primary_source_destroyed,
+                         data_device);
+
+      selection_source = meta_selection_source_wayland_new (source);
+    }
+  else
+    {
+      selection_source = g_object_new (META_TYPE_SELECTION_SOURCE_MEMORY, NULL);
+    }
+
+  set_selection_source (data_device, META_SELECTION_PRIMARY,
+                        selection_source);
+  g_object_unref (selection_source);
+}
+
+static void
+primary_device_set_selection (struct wl_client   *client,
+                              struct wl_resource *resource,
+                              struct wl_resource *source_resource,
+                              uint32_t            serial)
+{
+  MetaWaylandDataDevice *data_device = wl_resource_get_user_data (resource);
+  MetaWaylandSeat *seat = wl_container_of (data_device, seat, data_device);
+  MetaWaylandDataSource *source = NULL;
+
+  if (source_resource)
+    source = wl_resource_get_user_data (source_resource);
+
+  if (wl_resource_get_client (resource) !=
+      meta_wayland_keyboard_get_focus_client (seat->keyboard))
+    return;
+
+  meta_wayland_data_device_set_primary (data_device, source, serial);
+}
+
+static const struct gtk_primary_selection_device_interface primary_device_interface = {
+  primary_device_set_selection,
+  default_destructor,
+};
+
+static void
+destroy_data_source (struct wl_resource *resource)
+{
+  MetaWaylandDataSource *source = wl_resource_get_user_data (resource);
+
+  meta_wayland_data_source_set_resource (source, NULL);
+  g_object_unref (source);
+}
 
 static void
 create_data_source (struct wl_client *client,
@@ -976,7 +1869,23 @@ owner_changed_cb (MetaSelection         *selection,
   if (!focus_client)
     return;
 
-  if (selection_type == META_SELECTION_CLIPBOARD)
+  if (selection_type == META_SELECTION_PRIMARY)
+    {
+      wl_resource_for_each (data_device_resource, &data_device->primary_focus_resource_list)
+        {
+          struct wl_resource *offer = NULL;
+
+          if (new_owner)
+            {
+              offer = create_and_send_primary_offer (data_device,
+                                                     data_device_resource);
+            }
+
+          gtk_primary_selection_device_send_selection (data_device_resource,
+                                                       offer);
+        }
+    }
+  else if (selection_type == META_SELECTION_CLIPBOARD)
     {
       wl_resource_for_each (data_device_resource, &data_device->focus_resource_list)
         {
@@ -1026,12 +1935,72 @@ static const struct wl_data_device_manager_interface manager_interface = {
 };
 
 static void
+destroy_primary_source (struct wl_resource *resource)
+{
+  MetaWaylandDataSource *source = wl_resource_get_user_data (resource);
+
+  meta_wayland_data_source_set_resource (source, NULL);
+  g_object_unref (source);
+}
+
+static void
+primary_device_manager_create_source (struct wl_client   *client,
+                                      struct wl_resource *manager_resource,
+                                      guint32             id)
+{
+  struct wl_resource *source_resource;
+
+  source_resource =
+    wl_resource_create (client, &gtk_primary_selection_source_interface,
+                        wl_resource_get_version (manager_resource),
+                        id);
+  meta_wayland_data_source_primary_new (source_resource);
+}
+
+static void
+primary_device_manager_get_device (struct wl_client   *client,
+                                   struct wl_resource *manager_resource,
+                                   guint32             id,
+                                   struct wl_resource *seat_resource)
+{
+  MetaWaylandSeat *seat = wl_resource_get_user_data (seat_resource);
+  struct wl_resource *cr;
+
+  cr = wl_resource_create (client, &gtk_primary_selection_device_interface,
+                           wl_resource_get_version (manager_resource), id);
+  wl_resource_set_implementation (cr, &primary_device_interface,
+                                  &seat->data_device, unbind_resource);
+  wl_list_insert (&seat->data_device.primary_resource_list, wl_resource_get_link (cr));
+
+  ensure_owners_changed_handler_connected (&seat->data_device);
+}
+
+static const struct gtk_primary_selection_device_manager_interface primary_manager_interface = {
+  primary_device_manager_create_source,
+  primary_device_manager_get_device,
+  default_destructor,
+};
+
+static void
 bind_manager (struct wl_client *client,
               void *data, guint32 version, guint32 id)
 {
   struct wl_resource *resource;
   resource = wl_resource_create (client, &wl_data_device_manager_interface, version, id);
   wl_resource_set_implementation (resource, &manager_interface, NULL, NULL);
+}
+
+static void
+bind_primary_manager (struct wl_client *client,
+                      void             *data,
+                      uint32_t          version,
+                      uint32_t          id)
+{
+  struct wl_resource *resource;
+
+  resource = wl_resource_create (client, &gtk_primary_selection_device_manager_interface,
+                                 version, id);
+  wl_resource_set_implementation (resource, &primary_manager_interface, NULL, NULL);
 }
 
 void
@@ -1042,6 +2011,11 @@ meta_wayland_data_device_manager_init (MetaWaylandCompositor *compositor)
 			META_WL_DATA_DEVICE_MANAGER_VERSION,
 			NULL, bind_manager) == NULL)
     g_error ("Could not create data_device");
+
+  if (wl_global_create (compositor->wayland_display,
+			&gtk_primary_selection_device_manager_interface,
+			1, NULL, bind_primary_manager) == NULL)
+    g_error ("Could not create data_device");
 }
 
 void
@@ -1049,6 +2023,8 @@ meta_wayland_data_device_init (MetaWaylandDataDevice *data_device)
 {
   wl_list_init (&data_device->resource_list);
   wl_list_init (&data_device->focus_resource_list);
+  wl_list_init (&data_device->primary_resource_list);
+  wl_list_init (&data_device->primary_focus_resource_list);
 }
 
 static struct wl_resource *
@@ -1057,7 +2033,6 @@ create_and_send_clipboard_offer (MetaWaylandDataDevice *data_device,
 {
   MetaWaylandDataOffer *offer;
   MetaDisplay *display = meta_get_display ();
-  struct wl_resource *resource;
   GList *mimetypes, *l;
 
   mimetypes = meta_selection_get_mimetypes (meta_display_get_selection (display),
@@ -1065,17 +2040,57 @@ create_and_send_clipboard_offer (MetaWaylandDataDevice *data_device,
   if (!mimetypes)
     return NULL;
 
-  offer = meta_wayland_data_offer_new (META_SELECTION_CLIPBOARD, NULL, target);
-  resource = meta_wayland_data_offer_get_resource (offer);
+  offer = g_slice_new0 (MetaWaylandDataOffer);
+  offer->selection_type = META_SELECTION_CLIPBOARD;
+  offer->resource = wl_resource_create (wl_resource_get_client (target),
+                                        &wl_data_offer_interface,
+                                        wl_resource_get_version (target), 0);
+  wl_resource_set_implementation (offer->resource,
+                                  &data_offer_interface,
+                                  offer,
+                                  destroy_data_offer);
 
-  wl_data_device_send_data_offer (target, resource);
+  wl_data_device_send_data_offer (target, offer->resource);
 
   for (l = mimetypes; l; l = l->next)
-    wl_data_offer_send_offer (resource, l->data);
+    wl_data_offer_send_offer (offer->resource, l->data);
 
   g_list_free_full (mimetypes, g_free);
 
-  return resource;
+  return offer->resource;
+}
+
+static struct wl_resource *
+create_and_send_primary_offer (MetaWaylandDataDevice *data_device,
+                               struct wl_resource    *target)
+{
+  MetaWaylandDataOffer *offer;
+  MetaDisplay *display = meta_get_display ();
+  GList *mimetypes, *l;
+
+  mimetypes = meta_selection_get_mimetypes (meta_display_get_selection (display),
+                                            META_SELECTION_PRIMARY);
+  if (!mimetypes)
+    return NULL;
+
+  offer = g_slice_new0 (MetaWaylandDataOffer);
+  offer->selection_type = META_SELECTION_PRIMARY;
+  offer->resource = wl_resource_create (wl_resource_get_client (target),
+                                        &gtk_primary_selection_offer_interface,
+                                        wl_resource_get_version (target), 0);
+  wl_resource_set_implementation (offer->resource,
+                                  &primary_offer_interface,
+                                  offer,
+                                  destroy_primary_offer);
+
+  gtk_primary_selection_device_send_data_offer (target, offer->resource);
+
+  for (l = mimetypes; l; l = l->next)
+    gtk_primary_selection_offer_send_offer (offer->resource, l->data);
+
+  g_list_free_full (mimetypes, g_free);
+
+  return offer->resource;
 }
 
 void
@@ -1093,6 +2108,8 @@ meta_wayland_data_device_set_keyboard_focus (MetaWaylandDataDevice *data_device)
   data_device->focus_client = focus_client;
   move_resources (&data_device->resource_list,
                   &data_device->focus_resource_list);
+  move_resources (&data_device->primary_resource_list,
+                  &data_device->primary_focus_resource_list);
 
   if (!focus_client)
     return;
@@ -1107,6 +2124,18 @@ meta_wayland_data_device_set_keyboard_focus (MetaWaylandDataDevice *data_device)
 
       offer = create_and_send_clipboard_offer (data_device, data_device_resource);
       wl_data_device_send_selection (data_device_resource, offer);
+    }
+
+  move_resources_for_client (&data_device->primary_focus_resource_list,
+                             &data_device->primary_resource_list,
+                             focus_client);
+
+  wl_resource_for_each (data_device_resource, &data_device->primary_focus_resource_list)
+    {
+      struct wl_resource *offer;
+
+      offer = create_and_send_primary_offer (data_device, data_device_resource);
+      gtk_primary_selection_device_send_selection (data_device_resource, offer);
     }
 }
 
@@ -1124,8 +2153,72 @@ meta_wayland_data_device_get_current_grab (MetaWaylandDataDevice *data_device)
   return data_device->current_grab;
 }
 
-void
-meta_wayland_data_device_unset_dnd_selection (MetaWaylandDataDevice *data_device)
+gboolean
+meta_wayland_data_source_has_mime_type (const MetaWaylandDataSource *source,
+                                        const gchar                 *mime_type)
 {
-  unset_selection_source (data_device, META_SELECTION_DND);
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private ((MetaWaylandDataSource *)source);
+  gchar **p;
+
+  wl_array_for_each (p, &priv->mime_types)
+    {
+      if (g_strcmp0 (mime_type, *p) == 0)
+        return TRUE;
+    }
+
+  return FALSE;
+}
+
+static MetaWaylandDataSource *
+meta_wayland_data_source_new (struct wl_resource *resource)
+{
+  MetaWaylandDataSource *source =
+   g_object_new (META_TYPE_WAYLAND_DATA_SOURCE, NULL);
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+
+  meta_wayland_data_source_set_resource (source, resource);
+  wl_resource_set_implementation (resource, &data_source_interface,
+                                  source, destroy_data_source);
+
+  if (wl_resource_get_version (resource) < WL_DATA_SOURCE_ACTION_SINCE_VERSION)
+    {
+      priv->dnd_actions = priv->user_dnd_action =
+        WL_DATA_DEVICE_MANAGER_DND_ACTION_COPY;
+    }
+
+  return source;
+}
+
+static MetaWaylandDataSource *
+meta_wayland_data_source_primary_new (struct wl_resource *resource)
+{
+  MetaWaylandDataSource *source_primary =
+    g_object_new (META_TYPE_WAYLAND_DATA_SOURCE_PRIMARY, NULL);
+
+  meta_wayland_data_source_set_resource (source_primary, resource);
+  wl_resource_set_implementation (resource, &primary_source_interface,
+                                  source_primary, destroy_primary_source);
+
+  return source_primary;
+}
+
+gboolean
+meta_wayland_data_source_add_mime_type (MetaWaylandDataSource *source,
+                                        const gchar           *mime_type)
+{
+  MetaWaylandDataSourcePrivate *priv =
+    meta_wayland_data_source_get_instance_private (source);
+  gchar **pos;
+
+  pos = wl_array_add (&priv->mime_types, sizeof (*pos));
+
+  if (pos)
+    {
+      *pos = g_strdup (mime_type);
+      return *pos != NULL;
+    }
+
+  return FALSE;
 }
