+++ src/wayland/meta-wayland-data-device.h
@@ -28,20 +28,44 @@
 
 #include "clutter/clutter.h"
 #include "meta/meta-selection-source.h"
-#include "wayland/meta-wayland-data-offer.h"
-#include "wayland/meta-wayland-data-source.h"
 #include "wayland/meta-wayland-types.h"
 
 typedef struct _MetaWaylandDragGrab MetaWaylandDragGrab;
 typedef struct _MetaWaylandDataSourceFuncs MetaWaylandDataSourceFuncs;
 
+#define META_TYPE_WAYLAND_DATA_SOURCE (meta_wayland_data_source_get_type ())
+G_DECLARE_DERIVABLE_TYPE (MetaWaylandDataSource, meta_wayland_data_source,
+                          META, WAYLAND_DATA_SOURCE, GObject);
+
+struct _MetaWaylandDataSourceClass
+{
+  GObjectClass parent_class;
+
+  void (* send)    (MetaWaylandDataSource *source,
+                    const gchar           *mime_type,
+                    gint                   fd);
+  void (* target)  (MetaWaylandDataSource *source,
+                    const gchar           *mime_type);
+  void (* cancel)  (MetaWaylandDataSource *source);
+
+  void (* action)         (MetaWaylandDataSource *source,
+                           uint32_t               action);
+  void (* drop_performed) (MetaWaylandDataSource *source);
+  void (* drag_finished)  (MetaWaylandDataSource *source);
+};
+
 struct _MetaWaylandDataDevice
 {
   uint32_t selection_serial;
+  uint32_t primary_serial;
   MetaWaylandDataSource *selection_data_source;
   MetaWaylandDataSource *dnd_data_source;
+  MetaWaylandDataSource *primary_data_source;
+  struct wl_listener selection_data_source_listener;
   struct wl_list resource_list;
   struct wl_list focus_resource_list;
+  struct wl_list primary_resource_list;
+  struct wl_list primary_focus_resource_list;
   MetaWaylandDragGrab *current_grab;
   struct wl_client *focus_client;
 
@@ -67,7 +91,42 @@ void meta_wayland_data_device_set_dnd_source     (MetaWaylandDataDevice *data_de
 void meta_wayland_data_device_set_selection      (MetaWaylandDataDevice *data_device,
                                                   MetaWaylandDataSource *source,
                                                   guint32 serial);
-void     meta_wayland_data_device_unset_dnd_selection (MetaWaylandDataDevice *data_device);
+void meta_wayland_data_device_set_primary        (MetaWaylandDataDevice *data_device,
+                                                  MetaWaylandDataSource *source,
+                                                  guint32                serial);
+
+gboolean meta_wayland_data_source_add_mime_type  (MetaWaylandDataSource *source,
+                                                  const gchar           *mime_type);
+
+gboolean meta_wayland_data_source_has_mime_type  (const MetaWaylandDataSource *source,
+                                                  const gchar                 *mime_type);
+
+struct wl_array *
+         meta_wayland_data_source_get_mime_types (const MetaWaylandDataSource *source);
+
+gboolean meta_wayland_data_source_has_target     (MetaWaylandDataSource *source);
+
+void     meta_wayland_data_source_set_has_target (MetaWaylandDataSource *source,
+                                                  gboolean               has_target);
+
+void     meta_wayland_data_source_cancel         (MetaWaylandDataSource *source);
+
+void     meta_wayland_data_source_send           (MetaWaylandDataSource *source,
+                                                  const gchar           *mime_type,
+                                                  gint                   fd);
+
+void     meta_wayland_data_source_notify_finish  (MetaWaylandDataSource *source);
+
+uint32_t meta_wayland_data_source_get_actions        (MetaWaylandDataSource *source);
+uint32_t meta_wayland_data_source_get_user_action    (MetaWaylandDataSource *source);
+uint32_t meta_wayland_data_source_get_current_action (MetaWaylandDataSource *source);
+
+void     meta_wayland_data_source_set_actions        (MetaWaylandDataSource *source,
+                                                      uint32_t               dnd_actions);
+void     meta_wayland_data_source_set_user_action    (MetaWaylandDataSource *source,
+                                                      uint32_t               action);
+void     meta_wayland_data_source_set_current_action (MetaWaylandDataSource *source,
+                                                      uint32_t               action);
 
 const MetaWaylandDragDestFuncs *
          meta_wayland_data_device_get_drag_dest_funcs (void);
