+++ src/wayland/meta-wayland-surface.h
@@ -58,8 +58,6 @@ struct _MetaWaylandSurfaceRoleClass
                            MetaWaylandSurfaceState *pending);
   void (*apply_state) (MetaWaylandSurfaceRole  *surface_role,
                        MetaWaylandSurfaceState *pending);
-  void (*post_apply_state) (MetaWaylandSurfaceRole  *surface_role,
-                            MetaWaylandSurfaceState *pending);
   gboolean (*is_on_logical_monitor) (MetaWaylandSurfaceRole *surface_role,
                                      MetaLogicalMonitor     *logical_monitor);
   MetaWaylandSurface * (*get_toplevel) (MetaWaylandSurfaceRole *surface_role);
