+++ src/wayland/meta-xwayland-dnd.c
@@ -230,7 +230,7 @@ xdnd_send_position (MetaXWaylandDnd *dnd,
   XEvent xev = { 0 };
 
   user_action = meta_wayland_data_source_get_user_action (source);
-  meta_wayland_data_source_get_actions (source, &actions);
+  actions = meta_wayland_data_source_get_actions (source);
 
   if (user_action & actions)
     action = user_action;
@@ -367,7 +367,6 @@ transfer_cb (MetaSelection *selection,
     }
 
   g_output_stream_close (stream, NULL, NULL);
-  g_object_unref (stream);
 }
 
 static void
