+++ src/x11/meta-x11-selection.c
@@ -32,6 +32,13 @@
 #define UTF8_STRING_MIMETYPE "text/plain;charset=utf-8"
 #define STRING_MIMETYPE "text/plain"
 
+/* Set an arbitrary (although generous) threshold to determine whether a
+ * XFixesSelectionNotify corresponds to a XSetSelectionOwner from another
+ * client. The selection timestamp is not updated if the owner client is
+ * closed.
+ */
+#define SELECTION_CLEARED_BY_CLIENT(e) (e->timestamp - e->selection_timestamp < 50)
+
 static gboolean
 atom_to_selection_type (Display           *xdisplay,
                         Atom               selection,
@@ -335,7 +342,7 @@ meta_x11_selection_handle_xfixes_selection_notify (MetaX11Display *x11_display,
 
   if (event->owner == None && x11_display->selection.owners[selection_type])
     {
-      if (event->subtype == XFixesSetSelectionOwnerNotify)
+      if (SELECTION_CLEARED_BY_CLIENT (event))
         {
           MetaSelectionSource *source;
